ru.readmanga: moved Sorting enum into separate file
